Implement support for most item types/fields in 'add' command
The 'add' command reads a JSON file containing templates for
each item which specifies a list of sections, each with a list of fields.

It prompts for a value for each of the fields and saves the non-empty
ones to the new item's content.

The templates JSON file can be generated from a set of existing items
using the internal 'export-item-templates' command.

The initial item-templates.js file is generated from a set of items
exported to a vault by the Mac and iOS versions of 1Password with
all of the fields filled in.