Use an agent daemon to remember vault keys for a short time after unlock
Remember unlocked vault credentials for 45s after unlocking
the vault.

When 1pass is run, it spawns a daemon program using '1pass -agent'.
This provides an RPC interface to lock/unlock a vault given
a password and encrypt/decrypt item data for a given (vault, key name).

The client starts the agent if not already running and uses
it to unlock the vault if necessary.

The client still works directly with the vault contents but
when it needs to encrypt or decrypt item data, sends the data
to the vault which responds with the ciphertext/plaintext
for the item.