Change import/export commands to use the .1pif format used by the official apps
The 'export' and 'import' commands can now work with many items
at once and export to the .1pif format consisting of a
'<name>.1pif' directory containing a 'data.1pif' file.

The data.1pif file contains the unencrypted JSON for each item,
where the 'encrypted' field in a normal vault item is replaced
with an unencrypted 'secureContents' field.

If the item has attachments, the 1Password apps store these
in the 'attachments' dir inside the .1pif dir. This is currently
not supported.